+++ include/internal/catch_config.cpp
@@ -41,7 +41,6 @@ namespace Catch {
     }
 
     bool Config::listTests() const          { return m_data.listTests; }
-    bool Config::listTestNamesOnly() const  { return m_data.listTestNamesOnly; }
     bool Config::listTags() const           { return m_data.listTags; }
     bool Config::listReporters() const      { return m_data.listReporters; }